requirements.txt
langchain-community
langchain-openai
chromadb
langchain
langchain-core
python-dotenv

